voiceover/requirements.txt
pyaudio
keyboard
openai-whisper
numpy
python-espeak
sounddevice
scipy
pyyaml